First major commit
First big commit, at this point for simply one plug-in, an image-resizing plugin that resizes all layers of an image, multi-threaded using OpenMP and a partially complete custom variant of libquadmath, customized by self, and the whole thing is a work-in-progress that isn't ready for release yet.  Still clunky, has known issues.